requirements.txt
# LLM e modelos
transformers==4.41.1
accelerate==0.30.1
torch==2.3.0
sentence-transformers==2.7.0
ctransformers==0.2.27

# Embeddings e vetores
chromadb==0.5.0
faiss-cpu==1.8.0
langchain-community==0.2.3
langchain-core==0.2.3
langchain==0.2.3
langchainhub==0.1.17
langchain-openai==0.1.8
huggingface-hub==0.22.2

# CrewAI
crewai==0.30.10

# Manipulação de arquivos e ambiente
python-dotenv==1.0.1
unstructured==0.14.7
tiktoken==0.6.0

# Utilitários
pydantic==2.7.1
pypdf==4.2.0

# Para compatibilidade com Windows
uvicorn==0.30.1
fastapi==0.111.0

## Caso as versões acima não funcionem, tente as versões mais recentes, basta retirar o número da versão.